backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
twilio>=8.5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
aiohttp>=3.9.0  
